moved some things around so local clock has the timeout functionality, not default clock. added req latest for remote clock
